+++ RatingPrediction/RatingPrediction.cs
@@ -483,8 +483,8 @@ namespace RatingPrediction
 			// training data stats
 			int num_users = training_data.MaxUserID + 1; // TODO get correct data ...
 			int num_items = training_data.MaxItemID + 1;
-			long matrix_size = num_users * num_items;
-			long empty_size  = matrix_size - training_data.Count;
+			long matrix_size = (long) num_users * num_items;
+			long empty_size  = (long) matrix_size - training_data.Count;
 			Console.WriteLine("matrix {0} empty {1}", matrix_size, empty_size);
 			double sparsity = (double) 100L * empty_size / matrix_size;
 			Console.WriteLine(string.Format(ni, "training data: {0} users, {1} items, sparsity {2,0:0.#####}", num_users, num_items, sparsity));
@@ -492,8 +492,8 @@ namespace RatingPrediction
 			// test data stats
 			num_users = test_data.MaxUserID + 1;
 			num_items = test_data.MaxItemID + 1;
-			matrix_size = num_users * num_items;
-			empty_size  = matrix_size - test_data.Count;
+			matrix_size = (long) num_users * num_items;
+			empty_size  = (long) matrix_size - test_data.Count;
 			sparsity = (double) 100L * empty_size / matrix_size;
 			Console.WriteLine(string.Format(ni, "test data:     {0} users, {1} items, sparsity {2,0:0.#####}", num_users, num_items, sparsity));
 		}
